+++ src/common/js/global.js
@@ -2,9 +2,9 @@
  * pwix:bootbox/src/common/js/global.js
  */
 
-//console.log( 'pwix:bootbox/src/common/js/config.js defining globally exported pwixBootbox object' );
+//console.log( 'pwix:bootbox/src/common/js/config.js defining globally exported Bootbox object' );
 
-pwixBootbox = {
+Bootbox = {
     // internationalization
     i18n: {}
 };