+++ ChangeLog.md
@@ -2,24 +2,25 @@
 
 ## ChangeLog
 
-### 1.4.1-rc
+### 1.5.0-rc
 
     Release date: 
 
     - Remove superfluous call to pwixI18n.namespace() at client startup
+    - Rename globally exported pwixBootbox to Bootbox (bumping candidate version number)
 
 ### 1.4.0
 
     Release date: 2023- 6-22
 
     - Replace jquery-ui-dist requirement with standard jquery-ui
-    - Define pwixBootbox.i18n.namespace() to get the i18n namespace of the package (todo #2)
+    - Define Bootbox.i18n.namespace() to get the i18n namespace of the package (todo #2)
 
 ### 1.3.0
 
     Release date: 2023- 6- 9
 
-    - pwixBootbox.alert() now accepts a callback
+    - Bootbox.alert() now accepts a callback
 
 ### 1.2.0